pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "techpdfparser"
version = "0.1.0"
description = "A technical PDF parsing library"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.13",
    "pyyaml>=6.0",
    "pillow>=12.0",
    "pymupdf>=1.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
    "pytest-cov>=7.1",
    "ruff>=0.15",
    "mypy>=1.20",
]
gpu = [
    "mineru>=3.1",
    # Undeklarierte Transitiv-Deps von mineru 3.1.4 (per Import-Scan verifiziert):
    "shapely>=2.1",         # mineru.model.ocr.seal_crop
    "pyclipper>=1.4",       # mineru.model.utils.pytorchocr.postprocess.db_postprocess
    "omegaconf>=2.3",       # mineru.model.utils.tools.infer.pytorchocr_utility
    "albumentations>=2.0",  # mineru.model.mfr.unimernet (Formel-Pfad)
    "ftfy>=6.3",            # mineru.model.mfr.pp_formulanet_plus_m (Formel-Pfad)
    "transformers>=4.45,<5",
    "torch>=2.11",
    "torchvision>=0.26",
    "accelerate>=1.13",
    "beautifulsoup4>=4.14",
]
tables = [
    # Optional alternative table extractors (docling_table, tatr, qwen25vl_table).
    # The Qwen variant reuses transformers from [gpu]; tatr+docling_table need
    # tesseract for cell-level OCR (system binary: `apt install tesseract-ocr`).
    "docling-ibm-models>=3.0",
    "huggingface-hub>=0.26",
    "pytesseract>=0.3.10",
]

[tool.setuptools.packages.find]
include = ["extraction*"]
exclude = ["outputs*", "_archive*", "embedding*", "indexing*", "tests*", "tasks*"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
files = [
    "extraction",
]
check_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unreachable = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["yaml"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "mineru.*",
    "transformers.*",
    "torch.*",
    "bs4.*",
    "fitz.*",
    "pytesseract.*",
    "docling_ibm_models.*",
    "huggingface_hub.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["extraction/tests"]
addopts = "-m 'not integration'"
markers = [
    "integration: heavy end-to-end tests that require MinerU / GPU / extra models",
]
